Refactor after review comments p.1
