+++ templates/hotkeys.html
@@ -141,6 +141,39 @@
   </tr>
 </table>
 
+
+  <h2>Event Validation</h2>
+<table>
+  <tr>
+    <th style="width: 100px;">Hotkey</th>
+    <th style="width: 400px;">Action</th>
+  </tr>
+  <tr>
+    <td>1 to 9</td>
+    <td style="font-size: 16px;">Select answer per question</td>
+  </tr>
+  <tr>
+    <td>A</td>
+    <td style="font-size: 16px;">Switch to image mode</td>
+  </tr>
+  <tr>
+    <td>S</td>
+    <td style="font-size: 16px;">Switch to video mode</td>
+  </tr>
+  <tr>
+    <td>Z</td>
+    <td style="font-size: 16px;">Go to the previous video frame</td>
+  </tr>
+  <tr>
+    <td>X</td>
+    <td style="font-size: 16px;">Go to the next video frame</td>
+  </tr>
+  <tr>
+    <td>Esc</td>
+    <td style="font-size: 16px;">Exit from comment field</td>
+  </tr>
+</table>
+
 </div>
 </body>
 </html>
