+++ templates/how.html
@@ -105,6 +105,20 @@
         Adjust delay between frame change using keys 1, 2, 3 (short delay, middle delay, long delay). It should be balanced between speed and filtering accuracy.
     </p>
 
+    <h2>Event Validation</h2>
+    <p>
+        Use buttons <strong>1-9</strong> to iterate through answers for each question.
+    </p>
+    <p>
+        Click on <strong>Comment</strong> field to type the text and press <strong>Esc</strong> button to exit from it.
+    </p>
+    <p>
+        Switch between image/video mode by using keys <strong>A</strong> and <strong>S</strong>.
+    </p>
+    <p>
+        In video mode, use buttons <strong>Z</strong> and <strong>X</strong> for moving through frames backward and forward accordingly.
+    </p>
+
     </div>
 </body>
 </html>
